# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: AI Interview Platform API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /candidate/check-access:
    post:
      operationId: checkCandidateAccess
      tags: [candidate]
      summary: Check if candidate email has access to an interview
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CheckAccessBody"
      responses:
        "200":
          description: Access check result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CheckAccessResponse"
  /candidate/sessions:
    post:
      operationId: createSession
      tags: [candidate]
      summary: Start a new interview session
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateSessionBody"
      responses:
        "201":
          description: Created session
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Session"
  /candidate/sessions/{id}:
    get:
      operationId: getSession
      tags: [candidate]
      summary: Get session details
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Session
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Session"
  /candidate/sessions/{id}/next-question:
    post:
      operationId: getNextQuestion
      tags: [candidate]
      summary: Get next adaptive question for session
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Next question or completion signal
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/NextQuestionResponse"
  /candidate/sessions/{id}/answers:
    post:
      operationId: submitAnswer
      tags: [candidate]
      summary: Submit answer to current question
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SubmitAnswerBody"
      responses:
        "201":
          description: Answer submitted with AI evaluation
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AnswerWithEvaluation"
  /candidate/sessions/{id}/coding-questions:
    get:
      operationId: getCodingQuestions
      tags: [candidate]
      summary: Get coding questions for session
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Coding questions
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/CodingQuestion"
  /candidate/sessions/{id}/coding-answers:
    post:
      operationId: submitCodingAnswer
      tags: [candidate]
      summary: Submit coding answer
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SubmitCodingAnswerBody"
      responses:
        "201":
          description: Coding answer submitted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CodingAnswer"
  /candidate/sessions/{id}/complete:
    post:
      operationId: completeSession
      tags: [candidate]
      summary: Complete the interview session
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Session completed with results
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SessionWithResults"
  /candidate/sessions/{id}/results:
    get:
      operationId: getSessionResults
      tags: [candidate]
      summary: Get results for completed session
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Session results
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SessionWithResults"

  # Interview AI - TTS
...
components:
  schemas:
...
    JobProfileSkill:
      type: object
      properties:
        id:
          type: integer
        jobProfileId:
          type: integer
        skillName:
          type: string
        proficiencyLevel:
          type: string
          enum: [beginner, intermediate, advanced, expert]
        weightage:
          type: integer
      required:
        - id
        - jobProfileId
        - skillName
        - proficiencyLevel
        - weightage
...
    JobProfileWithSkills:
      type: object
      properties:
        id:
          type: integer
        employerId:
          type: integer
        title:
          type: string
        description:
          type: string
        createdAt:
          type: string
          format: date-time
        skills:
          type: array
          items:
            $ref: "#/components/schemas/JobProfileSkill"
      required:
        - id
        - employerId
        - title
        - description
        - createdAt
        - skills
...
    Interview:
      type: object
      properties:
        id:
          type: integer
        employerId:
          type: integer
        jobProfileId:
          type: integer
        title:
          type: string
        scheduledAt:
          type: ["string", "null"]
          format: date-time
        status:
          type: string
          enum: [draft, active, closed]
        difficulty:
          type: string
          enum: [easy, medium, hard]
        interviewerTone:
          type: string
          enum: [friendly, professional, strict, casual]
        numBehavioralQuestions:
          type: integer
        numTechnicalQuestions:
          type: integer
        numCodingQuestions:
          type: integer
        allowedCodingLanguages:
          type: string
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - employerId
        - jobProfileId
        - title
        - status
        - difficulty
        - interviewerTone
        - numBehavioralQuestions
        - numTechnicalQuestions
        - numCodingQuestions
        - allowedCodingLanguages
        - createdAt
...
    Session:
      type: object
      properties:
        id:
          type: integer
        interviewId:
          type: integer
        candidateEmail:
          type: string
        candidateName:
          type: string
        status:
          type: string
          enum: [in_progress, completed, abandoned]
        difficulty:
          type: string
          enum: [easy, medium, hard]
        interviewerTone:
          type: string
          enum: [friendly, professional, strict, casual]
        codingLanguage:
          type: ["string", "null"]
        overallScore:
          type: ["integer", "null"]
        currentQuestionIndex:
          type: integer
        totalQuestions:
          type: integer
        phase:
          type: string
          enum: [intro, behavioral, technical, coding, completed]
        startedAt:
          type: string
          format: date-time
        completedAt:
          type: ["string", "null"]
          format: date-time
      required:
        - id
        - interviewId
        - candidateEmail
        - candidateName
        - status
        - difficulty
        - interviewerTone
        - currentQuestionIndex
        - totalQuestions
        - phase
        - startedAt
    CreateSessionBody:
      type: object
      properties:
        interviewId:
          type: integer
        candidateEmail:
          type: string
        candidateName:
          type: string
        difficulty:
          type: string
          enum: [easy, medium, hard]
        interviewerTone:
          type: string
          enum: [friendly, professional, strict, casual]
        codingLanguage:
          type: string
      required:
        - interviewId
        - candidateEmail
        - candidateName
        - difficulty
        - interviewerTone
    Question:
      type: object
      properties:
        id:
          type: integer
        sessionId:
          type: integer
        type:
          type: string
          enum: [intro, behavioral, technical, coding]
        content:
          type: string
        difficulty:
          type: string
          enum: [easy, medium, hard]
        orderIndex:
          type: integer
        skillTag:
          type: ["string", "null"]
      required:
        - id
        - sessionId
        - type
        - content
        - difficulty
        - orderIndex
    NextQuestionResponse:
      type: object
      properties:
        question:
          $ref: "#/components/schemas/Question"
        sessionComplete:
          type: boolean
        phase:
          type: string
          enum: [intro, behavioral, technical, coding, completed]
        questionNumber:
          type: integer
        totalQuestions:
          type: integer
      required:
        - sessionComplete
        - phase
    SubmitAnswerBody:
      type: object
      properties:
        questionId:
          type: integer
        content:
          type: string
      required:
        - questionId
        - content
    AnswerWithEvaluation:
      type: object
      properties:
        id:
          type: integer
        sessionId:
          type: integer
        questionId:
          type: integer
        content:
          type: string
        score:
          type: ["integer", "null"]
        feedback:
          type: ["string", "null"]
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - sessionId
        - questionId
        - content
        - createdAt
    CodingQuestion:
      type: object
      properties:
        id:
          type: integer
        sessionId:
          type: integer
        language:
          type: string
        problem:
          type: string
        difficulty:
          type: string
          enum: [easy, medium, hard]
        orderIndex:
          type: integer
        exampleInput:
          type: ["string", "null"]
        exampleOutput:
          type: ["string", "null"]
      required:
        - id
        - sessionId
        - language
        - problem
        - difficulty
        - orderIndex
    SubmitCodingAnswerBody:
      type: object
      properties:
        codingQuestionId:
          type: integer
        code:
          type: string
        language:
          type: string
      required:
        - codingQuestionId
        - code
        - language
    CodingAnswer:
      type: object
      properties:
        id:
          type: integer
        sessionId:
          type: integer
        codingQuestionId:
          type: integer
        code:
          type: string
        language:
          type: string
        score:
          type: ["integer", "null"]
        feedback:
          type: ["string", "null"]
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - sessionId
        - codingQuestionId
        - code
        - language
        - createdAt
    SessionWithResults:
      type: object
      properties:
        session:
          $ref: "#/components/schemas/Session"
        answers:
          type: array
          items:
            $ref: "#/components/schemas/AnswerWithEvaluation"
        codingAnswers:
          type: array
          items:
            $ref: "#/components/schemas/CodingAnswer"
        skillScores:
          type: array
          items:
            $ref: "#/components/schemas/SkillScore"
        overallFeedback:
          type: ["string", "null"]
      required:
        - session
        - answers
        - codingAnswers
        - skillScores
    SkillScore:
      type: object
      properties:
        skillName:
          type: string
        score:
          type: integer
        weightage:
          type: integer
      required:
        - skillName
        - score
        - weightage
...
    CheckAccessBody:
      type: object
      properties:
        interviewId:
          type: integer
        email:
          type: string
      required:
        - interviewId
        - email
    CheckAccessResponse:
      type: object
      properties:
        hasAccess:
          type: boolean
        interview:
          $ref: "#/components/schemas/Interview"
        jobProfile:
          $ref: "#/components/schemas/JobProfileWithSkills"
      required:
        - hasAccess

    # TTS
...