# lib/api-zod/src/generated/types/jobProfileWithSkills.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * AI Interview Platform API
 * OpenAPI spec version: 0.1.0
 */
import type { JobProfileSkill } from "./jobProfileSkill";

export interface JobProfileWithSkills {
  id: number;
  employerId: number;
  title: string;
  description: string;
  createdAt: Date;
  skills: JobProfileSkill[];
}