# lib/api-zod/src/generated/types/nextQuestionResponse.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * AI Interview Platform API
 * OpenAPI spec version: 0.1.0
 */
import type { NextQuestionResponsePhase } from "./nextQuestionResponsePhase";
import type { Question } from "./question";

export interface NextQuestionResponse {
  question?: Question;
  sessionComplete: boolean;
  phase: NextQuestionResponsePhase;
  questionNumber?: number;
  totalQuestions?: number;
}